Add option to use sprockets derailleur file store from config
